# Conflicts:
#	src/main/resources/templates/fragments/form.html
#	src/main/resources/templates/fragments/header.html
#	src/main/resources/templates/layout/admin_layout.html
#	src/main/resources/templates/layout/aside-map.html
#	src/main/resources/templates/login.html
#	src/main/resources/templates/simulation/simulation-form.html
#	src/main/resources/templates/simulation/simulation.html

+++ src/main/java/gruner/huger/grunerhugel/model/formobjects/CreateSimulation.java
@@ -10,7 +10,7 @@ public class CreateSimulation {
     Date startDate;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     Date endDate;
-    int budget;
+    float budget;
     String tractorName;
     int numTractor;
     int numWorkers;
@@ -41,11 +41,11 @@ public class CreateSimulation {
         this.endDate = endDate;
     }
 
-    public int getBudget() {
+    public float getBudget() {
         return budget;
     }
 
-    public void setBudget(int budget) {
+    public void setBudget(float budget) {
         this.budget = budget;
     }
